These changes introduce a better permissions system.  It is mostly compatible with the old one and more compatible with FB6. HARDUID no longer has to be combined with SETUID and QUELL for them to work. HARDUID = Get permissions from the owner of the trigger SETUID = Get permissions from the trigger QUELL = Get permissions from the program
H and S flags only apply to programs UNLESS #/EXTPERMS flags are also set.
If # flag is set on a program, HARDUID and SETUID will get permissions from
the ACTION or TRIGGER that triggered them (or from the program that called it)

If Q and EXTPERMS are combined, and the owner of the program is QUELLed, the
program will lose permission to run as well.

SET and @set has been enhanced to allow EXTPERMS to be set on programs.