* interface.c: Fixed a compiler warning when UTF8_SUPPORT is undefined.
* Everything else:

config.h has been moved to config.h.in. config.h will not be automatically
generated if it's missing (i.e. new installs), otherwise the existing config.h
will be kept and configure will warn the user that they may be missing out on
newer features.

This change should put a stop to our test environment settings creeping into
the trunk builds that users rely on. Do NOT customize config.h.in and commit it
unless you are intentionally changing a default for all users.